tests/requirements.txt
yandex-taxi-testsuite[postgresql-binary] >= 0.1.35
websockets >= 11.0.3
pyjwt[crypto] >= 2.8.0
grpcio
grpcio-tools
psycopg2
bson
